benchmarks/appworld/pyproject.toml
[project]
# A worked example of running the AppWorld benchmark through an OpenAI Agents
# SDK harness, writing inference.net-format JSONL traces. Single-task
# end-to-end demo; not a leaderboard submission.
name = "halo-appworld-bench"
version = "0.1.0"
description = "AppWorld benchmark adapter for HALO's OpenAI Agents SDK harness."
requires-python = ">=3.11,<3.13"  # AppWorld pins to 3.11; SDK works through 3.12.
dependencies = [
    # AppWorld benchmark and its in-process FastAPI/SQLite world.
    # main is significantly ahead of the latest 0.1.3 PyPI release; only main
    # supports pydantic v2 (which the OpenAI Agents SDK requires).
    "appworld",
    # OpenAI Agents SDK. `tracing.py` plugs into the SDK's `TracingProcessor`
    # ABC and writes one JSONL span per line directly to a gzipped file.
    "openai-agents",
    # CLI / config conveniences shared with demo/openai-agents-sdk-demo.
    "typer",
    "rich",
    "python-dotenv",
    # Used by curate_and_upload.py to push traces to HuggingFace.
    "huggingface_hub",
]

[tool.uv.sources]
appworld = { git = "https://github.com/StonyBrookNLP/appworld", rev = "main" }
